Use the public annotations
The _annotations property was used to avoid creating the OrderedDict
when it wasn't necessary. Newer versions of Python don't need to use
OrderedDict, so the _annotations property has been removed. Instead,
we should just use the public .annotations property, which has been
around since Django 1.8. This has the potential to make some things
slower on older versions of Python or Django, but for compatibility
it seems like the right choice to me.