Determine the alias for the columns at the time of execution.
Add test to cover alias change in annotation.

Look up the column reference's table alias based on the CTE used.

Create failing test for CTE recursive query as subquery.

Fix alias issue when using a recursive CTE as a subquery.

Handle case in which the query is modifed after being used as a subquery.