Allow passing additional arguments to the `make_cte_queryset` function
Add `*args` and `**kwargs`arguments to the `With.recursive` method that will then be passed to the `make_cte_queryset` function.

This will allow further customization of the `make_cte_queryset` function, e.g. by passing `request.user` to it to filter by user directly in SQL or for any kind of dynamic filter.