Handle empty result sets in CTEs
If Django finds a CTE during SQL generation that will result in empty
results, it will throw an EmptyResultSet immediately. Because the CTEs
are compiled to SQL prior to the base query in the CTEQueryCompiler,
though, the base compiler is missing core information Django needs to
gracefully handle this situation, like `col_count` and `klass_info`.

To resolve this, in the case of EmptyResultSet being thrown by CTE
compilation, the base `as_sql` will be called as well before reraising.

The creation of the base queryset can not be move before the explain
behaviors in `generate_sql` or the generated `EXPLAIN` will throw an
error, so the try/except pattern is used instead.