Allow for empty result sets on left outer join
This explicitly passes in the `elide_empty` parameter to each CTE
compiler based on if the `join` used by it is an outer or inner join. By
passing in `elide_empty` as false, instead of the compiler raising an
`EmptyResultSet` error, it will instead alter the query to always return
no results by setting the where clause to something that is always
false.