Handle left outer joins in versions before 4.0
Before Django 4.0, the `elide_empty` argument on the compiler did not
exist. To ensure that left outer joins in these versions of Django are
not optimized away by the SqlCompile, essentially the same behavior is
added to the error handling in those versions of Django. This creates
a copy of the CTE query, forces the where to have a always-false
condition that the SqlCompiler cannot optimize away, and then builds
the SQL for that query instead.